# backend/docs/swagger.yaml
basePath: /api
definitions:
  middlewares.PermissionDenied:
    properties:
      error:
        type: string
    type: object
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  models.Permission:
    properties:
      oc_groups:
        type: boolean
      oc_users:
        type: boolean
      occtl:
        type: boolean
      see_server_logs:
        type: boolean
      statistic:
        type: boolean
      system:
        type: boolean
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_login:
        type: string
      permission:
        $ref: '#/definitions/models.Permission'
      uid:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - id
    - is_admin
    - last_login
    - permission
    - uid
    - username
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - page_size
    type: object
  user.ChangePasswordData:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
  user.ChangeStaffPassword:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  user.CreateStaffData:
    properties:
      password:
        type: string
      permission:
        $ref: '#/definitions/models.Permission'
      username:
        type: string
    required:
    - password
    - permission
    - username
    type: object
  user.StaffsResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - meta
    type: object
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /user/password:
    post:
      consumes:
      - application/json
      description: User Change Password
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: change user password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: User Change Password
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      description: Get user Profile
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Get user Profile
      tags:
      - User
  /user/staffs:
    get:
      consumes:
      - application/json
      description: List of Staffs
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.StaffsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: List of Staffs
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create Admin or Staff
      parameters:
      - description: user create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateStaffData'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Create Admin or Staff
      tags:
      - User
  /user/staffs/${id}/:
    delete:
      consumes:
      - application/json
      description: Remove Staff by given id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Remove Staff
      tags:
      - User
  /user/staffs/{id}/password:
    post:
      consumes:
      - application/json
      description: Change Staff Password by given id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: user new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangeStaffPassword'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Change Staff Password
      tags:
      - User
  /user/staffs/permissions/${id}/:
    put:
      consumes:
      - application/json
      description: Update Staff Permission by given id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: user permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Update Staff Permission
      tags:
      - User
swagger: "2.0"